Controlling-Lights-and-Monitoring-Temperature-Humidity-via-a-Web-Server-with-ESP32-Development-Board-
In this project, the goal is to turn on and off 3 different LEDs via a web server using the ESP32 development
board, and to dynamically read the temperature and humidity values of the environment using the DHT11
temperature and humidity sensor through the same server interface. All devices on the local Wi-Fi network
to which the ESP32 is connected can access and control the web server interface via the same IP address.